% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/VorTeq.R
\name{VorTeq_Tree}
\alias{VorTeq_Tree}
\title{Erosion Risk Potential}
\usage{
VorTeq_Tree(x)
}
\arguments{
\item{state}{input vector of VorTeq flows - comprising of 25 elements in the following order - # of RPXs running, process flow rate (gallons/min) through each of the 12 RPXs, motive flow rate (gallons/min) through each of the 12 RPXs}
}
\value{
output containes erosion potential and qualitative erosion risk comment
}
\examples{
VorTeq_Tree(state) # Returns erosion potential predicted by a decision tree model
}
